Fixes for local dev and bundle exec
